REF: implement replace_regex, remove unreachable branch in ObjectBlock.replace (#37696)
