+++ pandas/core/array_algos/replace.py
@@ -3,7 +3,7 @@ Methods used by Block.replace and related methods.
 """
 import operator
 import re
-from typing import Pattern, Union
+from typing import Optional, Pattern, Union
 
 import numpy as np
 
@@ -12,8 +12,10 @@ from pandas._typing import ArrayLike, Scalar
 from pandas.core.dtypes.common import (
     is_datetimelike_v_numeric,
     is_numeric_v_string_like,
+    is_re,
     is_scalar,
 )
+from pandas.core.dtypes.missing import isna
 
 
 def compare_or_regex_search(
@@ -87,3 +89,45 @@ def compare_or_regex_search(
 
     _check_comparison_types(result, a, b)
     return result
+
+
+def replace_regex(values: ArrayLike, rx: re.Pattern, value, mask: Optional[np.ndarray]):
+    """
+    Parameters
+    ----------
+    values : ArrayLike
+        Object dtype.
+    rx : re.Pattern
+    value : Any
+    mask : np.ndarray[bool], optional
+
+    Notes
+    -----
+    Alters values in-place.
+    """
+
+    # deal with replacing values with objects (strings) that match but
+    # whose replacement is not a string (numeric, nan, object)
+    if isna(value) or not isinstance(value, str):
+
+        def re_replacer(s):
+            if is_re(rx) and isinstance(s, str):
+                return value if rx.search(s) is not None else s
+            else:
+                return s
+
+    else:
+        # value is guaranteed to be a string here, s can be either a string
+        # or null if it's null it gets returned
+        def re_replacer(s):
+            if is_re(rx) and isinstance(s, str):
+                return rx.sub(value, s)
+            else:
+                return s
+
+    f = np.vectorize(re_replacer, otypes=[values.dtype])
+
+    if mask is None:
+        values[:] = f(values)
+    else:
+        values[mask] = f(values[mask])
